Moved validation back to init
