+++ cubequery/tasks/__init__.py
@@ -10,9 +10,11 @@ from urllib.request import Request, urlopen
 import datacube
 from jobtastic import JobtasticTask
 from shapely import wkt
+from shapely.geometry import shape, GeometryCollection
 
-from cubequery import get_config
+from cubequery import get_config, fetch_form_settings
 from cubequery.utils.s3_tools import S3Utils
+
 
 _http_headers = {"Content-Type": "application/json", "User-Agent": "cubequery-result"}
 
@@ -178,6 +180,105 @@ class CubeQueryTask(JobtasticTask):
 
     herd_avoidance_timeout = 60
     cache_duration = 60 * 60 * 24  # One day of seconds
+    
+    def standard_validation(self, args):
+        """
+        Validates conditions based upon the combination of the parameters provided.
+
+        Loads conditions set in input_conditions.json
+        
+        """
+        
+        _settings_json = fetch_form_settings()  
+        
+        if not _settings_json:          
+            with open('input_conditions.json') as res_json: 
+                _settings_json = json.load(res_json)  
+
+        keys = [k for k in _settings_json if k in args]
+
+        errors = []
+        
+        # Validates AOI
+        if 'aoi' in args:
+            errors = self.validate_standard_spatial_query(args['aoi'])
+        
+        # Validates information against input_conditions.json
+        for key in keys:
+            for d in _settings_json[key]:
+                if d['name'] == args[key]:
+                    for condition in d['conditions']:
+
+                        # Integer Range Validation
+                        if condition['type'] == 'int_range':
+                            for c in condition['id']:
+                                if c in args:                            
+                                    if len(condition['value'])==2:
+                                        if not (int(args[c]) >= condition['value'][0]) or not(int(args[c]) <= condition['value'][1]):
+                                            errors.append(create_error_message(condition))
+                                    else:
+                                        if not (int(args[c]) >= condition['value'][0]):
+                                            errors.append(create_error_message(condition))
+
+                        # Date Range Validation
+                        if condition['type'] == 'date_range':
+                            for c in condition['id']:
+                                if c in args:
+                                    if len(condition['value'])==2:
+                                        if not (args[c] >= condition['value'][0]) or not(args[c] <= condition['value'][1]):
+                                            errors.append(create_error_message(condition))
+                                    else:
+                                        if not (args[c] >= condition['value'][0]):
+                                            errors.append(create_error_message(condition))
+        
+        return errors
+
+    # TODO: Bounds conversion and sometimes spatial query dependent on product
+    def validate_standard_spatial_query(self, value):
+
+        errors = []
+
+        try:
+            parsed_polygon = wkt.loads(value)
+        except:
+            return [create_error_message({'id':'aoi', 'error_message':'Polygon could not be loaded', '_comment':'Polygon could not be loaded'})]
+
+        
+        '''
+        Returns validity of geometery (bool)
+        * Whole of Fiji = True
+        * Suva = True
+        '''
+        valid_geom = parsed_polygon.is_valid
+        if not valid_geom:
+            errors.append(create_error_message({'id':'aoi', 'error_message':'Geometry not a valid polygon', '_comment':'Geometry not a valid polygon'}))
+
+        '''
+        Returns area of polygon - About 1/4 of country ... 0.25 
+        * Whole of Fiji = 1.8662849915034905
+        * Suva = 0.017204474747948426
+        '''
+        area = parsed_polygon.area
+        if area > 0.25:
+            errors.append(create_error_message({'id':'aoi', 'error_message':'AOI area is too large', '_comment':'Size of polygon is too large'}))
+
+        '''
+        Returns bool for polygon inside Fiji
+        
+        For when antimeridian problem is resolved:
+        with open("TM_FIJI_BORDERS.geojson") as f:
+            features = json.load(f)["features"]
+            fiji_polygon = GeometryCollection([shape(feature["geometry"]).buffer(0) for feature in features])
+        '''
+            
+        fiji_polygon = wkt.loads('POLYGON((179.85466581578783 -15.573479020580137,177.13005644078854 -15.573479020580137,177.13005644078854 -19.40771171412428,179.85466581578783 -19.40771171412428,179.85466581578783 -15.573479020580137))')
+
+        contains = fiji_polygon.contains(parsed_polygon)
+        if contains == False:
+            errors.append(create_error_message({'id':'aoi', 'error_message':'AOI out of Fiji bounds', '_comment':'AOI is either completely or partially out of the Fiji bounds'}))
+
+        return errors
+
 
 
 def login_to_publisher():
@@ -250,3 +351,7 @@ def check_float(s):
         return True
     except ValueError:
         return False
+    
+
+def create_error_message(condition):
+    return {'Key':condition['id'], 'Error':condition['error_message'], 'Comment':condition['_comment']}
